[engine] refactor: migrate to a more universal filters model
In the early design, filters work like boolean predicates which iterate
over a sinlge column entry by entry. This model no longer fits into the
future design with event bus and PHLEX filtering. The new model filters
process entries from arbitrary columns in batch, which works for PHLEX
as well. As a bonus, the interface and implementation of Projection is
clearer and leaner.