[ui] refactor: move shared models into MonitorModeController
Clean up the mess in MonitorMode by moving all models shared by children
into a dedicated controller. Remove all references to MonitorMode in its
children.
At this stage, the impact of #2 is already mitigated to a great extent.
As the final step of the refactor, this commit eliminates all the direct
dependencies among MonitorMode and its children. While the encapsulation
of MonitorModeController is not yet ideal, the issue no longer presents
an obstacle to the development.
Closes #2.