[ui] feat: show schemaless options in the startup page
While verticalysis will attempt to infer the type of the input from its
file extension when no schema is selected, the info necessary for the
inference may not always be available. In such cases, users have to
explicitly indicate the source format, and schemaless options have to be
provided under such circumstances.